task(ModalBuy): css modules replaced by styled-components
