+++ components/base/Modal/index.tsx
@@ -1,3 +1,4 @@
+export { default as ModalBuy } from './ModalBuy';
 export { default as ModalMint } from './ModalMint';
 export { default as ModalShare } from './ModalShare';
 export { default as ModalWallet } from './ModalWallet';
