--- components/pages/NFT/ModalBuy/ModalBuy.module.scss
@@ -1,204 +1,0 @@
-@import 'style/colors.scss';
-
-.Background {
-  width: 100%;
-  position: fixed;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 2000;
-  box-sizing: border-box;
-}
-
-.Container {
-  width: 550px;
-  border-radius: 25px;
-  background-color: black;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  align-items: center;
-  padding: 70px 50px;
-  position: relative;
-}
-
-.Label {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: white;
-  background-color: $primary;
-  font-family: 'Airbnb Cereal App Medium';
-  font-size: 10px;
-  text-transform: uppercase;
-  border-radius: 25px;
-  padding: 6px 14px;
-}
-
-.Title {
-  font-family: 'Airbnb Cereal App Bold';
-  color: white;
-  font-size: 38px;
-  line-height: 42px;
-  text-align: center;
-  margin-top: 24px;
-}
-
-.Text {
-  color: white;
-  font-size: 14px;
-  line-height: 20px;
-  font-family: 'Airbnb Cereal App Medium';
-  margin-top: 16px;
-  text-align: center;
-  max-width: 400px;
-}
-
-.Close {
-  display: flex;
-  position: absolute;
-  z-index: 130;
-  fill: white;
-  top: 28px;
-  right: 28px;
-  width: 18px;
-  cursor: pointer;
-}
-
-.Button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: black;
-  background-color: white;
-  font-family: 'Airbnb Cereal App Medium';
-  font-size: 12px;
-  border-radius: 25px;
-  padding: 12px 60px;
-  cursor: pointer;
-  margin-top: 40px;
-  transition: all 0.4s cubic-bezier(0.25, 1, 0.5, 1);
-
-  &:hover {
-    color: white;
-    background-color: $primary;
-  }
-}
-
-.CheckMark {
-  fill: $primary;
-  width: 22px;
-  margin: 24px 0px;
-}
-
-.Status {
-  display: flex;
-  flex-direction: column;
-  align-self: center;
-  margin-top: 24px;
-  text-align: left;
-
-  span {
-    color: white;
-    margin-bottom: 8px;
-    font-family: 'Airbnb Cereal App Medium';
-    font-size: 14px;
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    text-align: left;
-  }
-}
-
-.Step {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  text-align: left;
-}
-
-.Error {
-  font-family: 'Airbnb Cereal App Medium';
-  font-size: 14px;
-  color: red;
-  align-self: center;
-  margin-top: 24px;
-  text-align: center;
-}
-
-.CheckMark {
-  fill: $primary;
-  width: 12px;
-  margin-left: 6px;
-}
-
-.QR {
-  margin-top: 40px;
-}
-
-.Loading {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 40px;
-  margin-bottom: 20px;
-}
-
-.Dot {
-  width: 6px;
-  height: 6px;
-  border-radius: 50%;
-  background-color: white;
-  margin: 0px 3px;
-  animation: load 1s cubic-bezier(0.25, 1, 0.5, 1) infinite;
-}
-
-.Dot:nth-child(2) {
-  animation-delay: 100ms;
-}
-
-.Dot:nth-child(3) {
-  animation-delay: 200ms;
-}
-
-@keyframes load {
-  0% {
-    transform: translateY(0px);
-  }
-  50% {
-    transform: translateY(-10px);
-  }
-  100% {
-    transform: translateY(0px);
-  }
-}
-
-@media (max-width: 720px) {
-  .Container {
-    width: 100%;
-    margin: 0px 24px;
-    padding: 52px 32px;
-  }
-
-  .Title {
-    font-size: 34px;
-    line-height: 38px;
-    margin-top: 24px;
-  }
-
-  .Text {
-    font-size: 13px;
-    line-height: 18px;
-    margin-top: 12px;
-    text-align: center;
-    max-width: 85%;
-  }
-
-  .Button {
-    font-size: 11px;
-    padding: 10px 28px;
-    margin-top: 32px;
-  }
-}
