--- components/pages/NFT/ModalBuy/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ModalBuy';
